refactor(app): split the three giant components by what they render (STR-2)
`SkillDetail.tsx` (2,245), `App.tsx` (1,476) and `FileTree.tsx` (1,738) each
kept a pile of private components or effects that never needed to be in the
same file. Pure moves, verified line-by-line against HEAD: outside the import
blocks, SkillDetail and App gained nothing but the new import lines and three
relocated comments, and FileTree gained only the two hook call sites.

  SkillDetail.tsx  2,245 -> 1,221 + 10 files, largest 273
  App.tsx          1,476 ->   911 +  4 files, largest 230
  FileTree.tsx     1,738 -> 1,320 +  3 files, largest 272

`teamshare/skill-detail/` takes the eight private components the detail pane
rendered — `ModalShell`, `UsageBoundary`, `MetaRow`, `VersionHistory`,
`ShareSheet`, `ConflictBar`, `PublishVersionSheet`, `ForkSheet`, `DiffSheet` —
plus the one date helper two of them share. The `DiffRenderer` lazy boundary
moves with `DiffSheet`, which is its only user.

`app/` takes what the shell owns rather than the screen: `shell-hooks` (tray
locale sync, the ⌘, menu listener, daemon live status, webview and terminal
shortcuts), `MainContent` (the centre column, carrying the three lazy panes
only it renders), `chrome` (terminal toggle, header tab, resize handle) and
the find-bar/zoom micro-store the shortcuts and the webview surface both read.
`AppContent` stays behind.

`workspace/file-tree/` takes `flatten` (filter + flatten, pure, now testable
without mounting the tree), `use-file-tree-keyboard` (200 lines of key
handling) and `use-os-file-drop` (the Tauri drag-drop listeners). Both hooks
keep the original dependency arrays verbatim, so memoisation and
re-subscription are unchanged; refs and setState identities passed in as
parameters are exactly the ones eslint can no longer prove stable, hence the
two disable comments with that reason written down.

Also reattaches four comments that had drifted from what they document (one
of them, the webview-shortcut block, was already misplaced on main).

Verification: typecheck clean with noUnusedLocals on, lint 0 errors and the
same 48 warnings as HEAD, unit suite 488 files / 3,244 tests green.
