+++ packages/app/src/components/teamshare/SkillDetail.tsx
@@ -1,1057 +1,33 @@
 import * as React from 'react'
 import { Suspense, lazy } from 'react'
 import { useTranslation } from 'react-i18next'
-import {
-  Sparkles,
-  Save,
-  Loader2,
-  Download,
-  Share2,
-  Trash2,
-  Archive,
-  AlertTriangle,
-  Copy,
-  Upload,
-} from 'lucide-react'
+import { Sparkles, Save, Loader2, Download, Share2, Trash2, Archive, AlertTriangle, Copy } from 'lucide-react'
 import { toast } from 'sonner'
 import { Button } from '@/components/ui/button'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog'
+import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { cn } from '@/lib/utils'
 import { encodeWorkspaceId, putDaemonSkill } from '@/lib/daemon-local-client'
-import {
-  useTeamShareBrowserStore,
-  isSkillDirtyConflict,
-  SkillDiscardIncompleteError,
-  SkillSlugTakenError,
-  SkillPublishedRefreshError,
-  SkillMutationRefreshError,
-  SkillRuntimeRefreshError,
-  StaleTeamSkillPublishError,
-  StaleDirtySkillPublishError,
-  type TeamSkillItem,
-  type TeamSkillFileDiff,
-  type TeamSkillDraftMetadata,
-  type DraftRecoveryRecord,
-} from '@/stores/team-share-browser'
+import { useTeamShareBrowserStore, isSkillDirtyConflict, SkillDiscardIncompleteError, SkillSlugTakenError, SkillPublishedRefreshError, SkillMutationRefreshError, SkillRuntimeRefreshError, StaleTeamSkillPublishError, StaleDirtySkillPublishError, type TeamSkillFileDiff, type DraftRecoveryRecord } from '@/stores/team-share-browser'
 import { useCurrentTeamStore } from '@/stores/current-team'
 import { getBackend } from '@/lib/backend/provider'
-import {
-  TEAM_SKILL_CATEGORIES,
-  type TeamSkillCategory,
-  type TeamSkillVersion,
-} from '@/lib/backend/cloud-api/team-skills'
+import { type TeamSkillCategory, type TeamSkillVersion } from '@/lib/backend/cloud-api/team-skills'
 import { useIsDark } from './use-is-dark'
 import { resolveAgentDevicePresenceSync } from '@/lib/agent-device-reachability'
 import { useActorPresenceStore } from '@/stores/actor-presence-store'
 import { useEffectiveWorkspacePath } from '@/lib/effective-workspace'
 import { toastSkillMutationRefreshFailed } from './skillMutationRefreshToast'
+import { ConflictBar } from './skill-detail/ConflictBar'
+import { DiffSheet } from './skill-detail/DiffSheet'
+import { ForkSheet } from './skill-detail/ForkSheet'
+import { MetaRow } from './skill-detail/MetaRow'
+import { PublishVersionSheet } from './skill-detail/PublishVersionSheet'
+import { ShareSheet } from './skill-detail/ShareSheet'
+import { UsageBoundary } from './skill-detail/UsageBoundary'
+import { VersionHistory } from './skill-detail/VersionHistory'
 
 type SkillConfirmAction = 'delete' | 'uninstall'
 
 const CodeEditor = lazy(() => import('@/components/editors/CodeEditor'))
-const DiffRenderer = lazy(() =>
-  import('@/components/diff/DiffRenderer').then((m) => ({ default: m.DiffRenderer })),
-)
-
-/** Shared chrome for the conflict flow's dialogs. */
-function ModalShell({
-  title,
-  hint,
-  children,
-  footer,
-  onClose,
-  wide,
-}: {
-  title: string
-  hint?: string
-  children: React.ReactNode
-  footer: React.ReactNode
-  onClose: () => void
-  wide?: boolean
-}) {
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-black/30 p-4"
-      onClick={onClose}
-    >
-      <div
-        role="dialog"
-        aria-modal="true"
-        onClick={(e) => e.stopPropagation()}
-        className={cn(
-          'flex max-h-[85vh] w-full flex-col overflow-hidden rounded-[14px] border border-border bg-paper shadow-lg',
-          wide ? 'max-w-3xl' : 'max-w-lg',
-        )}
-      >
-        <div className="border-b border-border px-5 py-3">
-          <h2 className="text-[15px] font-bold text-foreground">{title}</h2>
-          {hint && <p className="mt-1 text-[12px] text-muted-foreground">{hint}</p>}
-        </div>
-        <div className="min-h-0 flex-1 space-y-3 overflow-y-auto px-5 py-4">{children}</div>
-        <div className="flex justify-end gap-2 border-t border-border px-5 py-3">{footer}</div>
-      </div>
-    </div>
-  )
-}
-
-function UsageBoundary({ item }: { item: TeamSkillItem }) {
-  const { t } = useTranslation()
-  if (!item.whenToUse && !item.whenNotToUse) return null
-  return (
-    <div className="grid grid-cols-1 gap-3 border-b border-border px-5 py-4 sm:grid-cols-2">
-      <section>
-        <h3 className="mb-1.5 text-[11px] font-semibold uppercase tracking-wide text-faint">
-          {t('teamShare.skillWhenToUse', 'When to use')}
-        </h3>
-        <p className="whitespace-pre-wrap text-[13px] leading-relaxed text-foreground">
-          {item.whenToUse || t('teamShare.skillFieldEmpty', '—')}
-        </p>
-      </section>
-      <section>
-        <h3 className="mb-1.5 text-[11px] font-semibold uppercase tracking-wide text-faint">
-          {t('teamShare.skillWhenNotToUse', 'When not to use')}
-        </h3>
-        <p className="whitespace-pre-wrap text-[13px] leading-relaxed text-foreground">
-          {item.whenNotToUse || t('teamShare.skillFieldEmpty', '—')}
-        </p>
-      </section>
-    </div>
-  )
-}
-
-function formatShortDate(iso: string | null | undefined): string | null {
-  if (!iso) return null
-  const d = new Date(iso)
-  if (Number.isNaN(d.getTime())) return null
-  return d.toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' })
-}
-
-function MetaRow({ item, ownerLabel }: { item: TeamSkillItem; ownerLabel: string | null }) {
-  const { t } = useTranslation()
-  const primary: string[] = []
-  if (item.kind === 'personal' && item.personalSourceLabel) primary.push(item.personalSourceLabel)
-  if (item.category) primary.push(item.category)
-  if (item.latestVersion) primary.push(`v${item.latestVersion}`)
-  if (item.installed && item.installedVersion && item.installedVersion !== item.latestVersion) {
-    primary.push(t('teamShare.skillInstalledVersion', 'installed v{{v}}', { v: item.installedVersion }))
-  }
-  if (item.requires?.length) {
-    primary.push(t('teamShare.skillRequiresList', 'requires {{list}}', { list: item.requires.join(', ') }))
-  }
-
-  const secondary: string[] = []
-  if (ownerLabel) {
-    secondary.push(t('teamShare.skillOwner', 'Owner · {{name}}', { name: ownerLabel }))
-  }
-  const updated = formatShortDate(item.updatedAt)
-  if (updated) {
-    secondary.push(t('teamShare.skillUpdated', 'Updated {{date}}', { date: updated }))
-  }
-
-  if (!primary.length && !secondary.length) return null
-  return (
-    <div className="space-y-1 border-b border-border px-5 py-2 text-[12px] text-muted-foreground">
-      {primary.length > 0 && <div>{primary.join(' · ')}</div>}
-      {secondary.length > 0 && <div className="text-[11px] text-faint">{secondary.join(' · ')}</div>}
-    </div>
-  )
-}
-
-function VersionHistory({
-  versions,
-  loading,
-  installedVersion,
-  canRevert,
-  reverting,
-  onRevert,
-}: {
-  versions: TeamSkillVersion[]
-  loading: boolean
-  installedVersion: number | null
-  canRevert: boolean
-  reverting: boolean
-  onRevert: (version: number) => void
-}) {
-  const { t } = useTranslation()
-  if (loading) {
-    return (
-      <div className="flex items-center gap-2 border-b border-border px-5 py-3 text-[12px] text-muted-foreground">
-        <Loader2 className="h-3.5 w-3.5 animate-spin" />
-        {t('common.loading', 'Loading…')}
-      </div>
-    )
-  }
-  if (!versions.length) return null
-
-  const sorted = [...versions].sort((a, b) => b.version - a.version)
-  const latest = sorted[0]
-  const older = sorted.slice(1, 4)
-
-  return (
-    <div className="border-b border-border px-5 py-4">
-      <h3 className="mb-2 text-[11px] font-semibold uppercase tracking-wide text-faint">
-        {t('teamShare.skillVersions', 'Versions')}
-      </h3>
-      <div className="rounded-[8px] border border-border-soft bg-paper/60 px-3 py-2.5">
-        <div className="flex items-baseline justify-between gap-2">
-          <span className="font-mono text-[12px] font-semibold text-foreground">
-            v{latest.version}
-            {installedVersion === latest.version && (
-              <span className="ml-2 font-sans text-[10.5px] font-medium text-faint">
-                {t('teamShare.skillInstalled', 'Installed')}
-              </span>
-            )}
-          </span>
-          <span className="shrink-0 font-mono text-[10.5px] text-faint">
-            {formatShortDate(latest.createdAt) ?? ''}
-          </span>
-        </div>
-        <p className="mt-1 whitespace-pre-wrap text-[12.5px] leading-relaxed text-ink-2">
-          {latest.changelog || t('teamShare.skillFieldEmpty', '—')}
-        </p>
-        {latest.publishedFromVersion != null && (
-          <p className="mt-1 font-mono text-[10.5px] text-faint">
-            {t('teamShare.skillPublishedFrom', 'Based on v{{v}}', { v: latest.publishedFromVersion })}
-          </p>
-        )}
-      </div>
-      {older.length > 0 && (
-        <ul className="mt-2 space-y-1.5">
-          {older.map((v) => (
-            <li key={v.version} className="flex items-baseline gap-2 text-[11.5px] text-muted-foreground">
-              <span className="shrink-0 font-mono text-faint">v{v.version}</span>
-              <span className="min-w-0 flex-1 truncate">
-                {v.changelog || t('teamShare.skillFieldEmpty', '—')}
-              </span>
-              {/*
-                The undo for a bad publish. Under auto-follow a broken version
-                is on every member's disk within one tick, so "publish a fix"
-                is not a remedy the author can always reach for — they may not
-                have the old bytes any more. This re-publishes them.
-              */}
-              {canRevert && (
-                <button
-                  type="button"
-                  disabled={reverting}
-                  onClick={() => onRevert(v.version)}
-                  className="shrink-0 text-[11px] text-muted-foreground underline-offset-2 hover:text-foreground hover:underline disabled:opacity-40"
-                >
-                  {t('teamShare.skillRevertTo', 'Restore')}
-                </button>
-              )}
-            </li>
-          ))}
-        </ul>
-      )}
-    </div>
-  )
-}
-
-function ShareSheet({
-  item,
-  open,
-  onClose,
-  onSubmit,
-  busy,
-  takenSlugs,
-}: {
-  item: TeamSkillItem
-  open: boolean
-  onClose: () => void
-  onSubmit: (input: {
-    slug: string
-    summary: string
-    category: TeamSkillCategory
-    whenToUse: string
-    whenNotToUse: string
-    changelog: string
-  }) => Promise<void>
-  busy: boolean
-  /** Names the team registry already owns. */
-  takenSlugs: ReadonlySet<string>
-}) {
-  const { t } = useTranslation()
-  const [slug, setSlug] = React.useState(item.slug)
-  const [summary, setSummary] = React.useState(item.summary ?? '')
-  const [category, setCategory] = React.useState<TeamSkillCategory>(
-    (TEAM_SKILL_CATEGORIES.includes(item.category as TeamSkillCategory)
-      ? item.category
-      : 'general') as TeamSkillCategory,
-  )
-  const [whenToUse, setWhenToUse] = React.useState(item.whenToUse ?? '')
-  const [whenNotToUse, setWhenNotToUse] = React.useState(item.whenNotToUse ?? '')
-  const [changelog, setChangelog] = React.useState('v1: shared from personal skill')
-
-  React.useEffect(() => {
-    if (!open) return
-    setSlug(item.slug)
-    setSummary(item.summary ?? '')
-    setCategory(
-      (TEAM_SKILL_CATEGORIES.includes(item.category as TeamSkillCategory)
-        ? item.category
-        : 'general') as TeamSkillCategory,
-    )
-    setWhenToUse(item.whenToUse ?? '')
-    setWhenNotToUse(item.whenNotToUse ?? '')
-    setChangelog('v1: shared from personal skill')
-  }, [open, item])
-
-  if (!open) return null
-
-  // The registry rejects a duplicate name, but only after the package has been
-  // uploaded. Everything needed to know that is already in this list, so say it
-  // while they are still typing.
-  const slugTaken = takenSlugs.has(slug.trim())
-
-  // Guidance fields are not gates. Requiring them mostly bought placeholder
-  // text, which reads as guidance without being any — worse than a blank the
-  // author can come back and fill in.
-  const canSubmit = slug.trim() && !slugTaken && summary.trim() && changelog.trim() && !busy
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/30 p-4">
-      <div
-        role="dialog"
-        aria-modal="true"
-        className="flex max-h-[85vh] w-full max-w-lg flex-col overflow-hidden rounded-[14px] border border-border bg-paper shadow-lg"
-      >
-        <div className="border-b border-border px-5 py-3">
-          <h2 className="text-[15px] font-bold text-foreground">
-            {t('teamShare.skillShareTitle', 'Share to team')}
-          </h2>
-          <p className="mt-1 text-[12px] text-muted-foreground">
-            {t(
-              'teamShare.skillShareHint',
-              'Publishes a copy to the team registry and installs it for you. Your personal folder stays on disk.',
-            )}
-          </p>
-        </div>
-        <div className="min-h-0 flex-1 space-y-3 overflow-y-auto px-5 py-4">
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillShareSlug', 'Slug')}
-            </span>
-            <input
-              value={slug}
-              onChange={(e) => setSlug(e.target.value)}
-              aria-invalid={slugTaken || undefined}
-              className={cn(
-                'w-full rounded-[8px] border bg-background px-3 py-2 font-mono text-[13px] outline-none',
-                slugTaken ? 'border-foreground' : 'border-border focus:border-coral/60',
-              )}
-            />
-            {slugTaken && (
-              <p className="text-[12px] leading-relaxed text-foreground">
-                {t(
-                  'teamShare.skillShareSlugTaken',
-                  'The team already has a skill called {{slug}}. Pick another name, or publish a new version of the existing one.',
-                  { slug: slug.trim() },
-                )}
-              </p>
-            )}
-          </label>
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillShareSummary', 'Summary')}
-            </span>
-            <input
-              value={summary}
-              onChange={(e) => setSummary(e.target.value)}
-              maxLength={200}
-              className="w-full rounded-[8px] border border-border bg-background px-3 py-2 text-[13px] outline-none focus:border-coral/60"
-            />
-          </label>
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillShareCategory', 'Category')}
-            </span>
-            <Select value={category} onValueChange={(v) => setCategory(v as TeamSkillCategory)}>
-              <SelectTrigger className="h-auto w-full rounded-[8px] border-border bg-background px-3 py-2 text-[13px] shadow-none">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                {TEAM_SKILL_CATEGORIES.map((c) => (
-                  <SelectItem key={c} value={c} className="text-[13px]">
-                    {c}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </label>
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillWhenToUse', 'When to use')}
-            </span>
-            <textarea
-              value={whenToUse}
-              onChange={(e) => setWhenToUse(e.target.value)}
-              rows={3}
-              className="w-full rounded-[8px] border border-border bg-background px-3 py-2 text-[13px] outline-none focus:border-coral/60"
-            />
-          </label>
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillWhenNotToUse', 'When not to use')}
-            </span>
-            <textarea
-              value={whenNotToUse}
-              onChange={(e) => setWhenNotToUse(e.target.value)}
-              rows={3}
-              className="w-full rounded-[8px] border border-border bg-background px-3 py-2 text-[13px] outline-none focus:border-coral/60"
-            />
-          </label>
-          <label className="block space-y-1">
-            <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-              {t('teamShare.skillShareChangelog', 'Changelog')}
-            </span>
-            <textarea
-              value={changelog}
-              onChange={(e) => setChangelog(e.target.value)}
-              rows={2}
-              className="w-full rounded-[8px] border border-border bg-background px-3 py-2 text-[13px] outline-none focus:border-coral/60"
-            />
-          </label>
-        </div>
-        <div className="flex justify-end gap-2 border-t border-border px-5 py-3">
-          <Button type="button" variant="ghost" onClick={onClose} disabled={busy} className="h-8 text-[13px]">
-            {t('common.cancel', 'Cancel')}
-          </Button>
-          <Button
-            type="button"
-            disabled={!canSubmit}
-            onClick={() =>
-              void onSubmit({
-                slug: slug.trim(),
-                summary: summary.trim(),
-                category,
-                whenToUse: whenToUse.trim(),
-                whenNotToUse: whenNotToUse.trim(),
-                changelog: changelog.trim(),
-              })
-            }
-            className="h-8 gap-1.5 bg-coral text-[13px] font-semibold text-white hover:bg-coral/90 disabled:opacity-40"
-          >
-            {busy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Share2 className="h-3.5 w-3.5" />}
-            {t('teamShare.skillShareSubmit', 'Share & install')}
-          </Button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-/**
- * Shown when auto-follow has stopped because the pack was edited locally.
- *
- * This is the only place in the skills UI that asks the user for anything, so
- * it is the only place that gets emphasis. A version that is merely behind
- * resolves itself within a reconcile tick and is reported as quiet meta text —
- * treating a self-healing state as an alert is how alerts stop meaning
- * anything.
- *
- * Naming the changed files is load-bearing rather than decorative. The user's
- * first question is "did I change this?", and a list of paths answers it
- * outright; "this skill has local modifications" leaves them opening a diff to
- * find out whether they care.
- */
-function ConflictBar({
-  modified,
-  deleted,
-  added,
-  installedVersion,
-  latestVersion,
-  busy,
-  canPublish,
-  isStaleDirty,
-  source,
-  onViewDiff,
-  onPublish,
-  onFork,
-  onDiscard,
-  onRebaseOnLatest,
-}: {
-  modified: string[]
-  deleted: string[]
-  added: string[]
-  installedVersion: number | null
-  latestVersion: number | null
-  busy: boolean
-  canPublish: boolean
-  isStaleDirty: boolean
-  source: 'hosted-agent' | 'member'
-  onViewDiff: () => void
-  onPublish: () => void
-  onFork: () => void
-  onDiscard: () => void
-  onRebaseOnLatest: () => void
-}) {
-  const { t } = useTranslation()
-  const changed = [
-    modified.length
-      ? t('teamShare.skillConflictModified', '{{files}} changed', { files: modified.join('、') })
-      : null,
-    deleted.length
-      ? t('teamShare.skillConflictDeleted', '{{files}} deleted', { files: deleted.join('、') })
-      : null,
-    added.length
-      ? t('teamShare.skillConflictAdded', '{{files}} added', { files: added.join('、') })
-      : null,
-  ]
-    .filter(Boolean)
-    .join(' · ')
-
-  return (
-    <div className="border-b border-border px-5 py-3">
-      <div className="rounded-[8px] border border-border border-l-2 border-l-faint bg-paper px-4 py-3">
-        <div className="flex items-baseline justify-between gap-3">
-          <span className="text-[13px] font-semibold text-foreground">
-            {isStaleDirty
-              ? t('teamShare.skillStaleConflictTitle', 'Your draft is behind the team')
-              : t('teamShare.skillConflictTitle', 'Local changes — updates paused')}
-          </span>
-          {latestVersion != null && installedVersion != null && (
-            <span className="shrink-0 font-mono text-[11px] text-faint">
-              {t('teamShare.skillConflictVersions', 'team v{{latest}} · you v{{installed}}', {
-                latest: latestVersion,
-                installed: installedVersion,
-              })}
-            </span>
-          )}
-        </div>
-        {isStaleDirty && (
-          <p className="mt-1 text-[12px] leading-relaxed text-muted-foreground">
-            {t(
-              'teamShare.skillStaleConflictBody',
-              'Your edits are based on v{{base}}. The team is on v{{latest}} — someone else may have published while you were editing. Drafts stay on this device until you publish.',
-              { base: installedVersion, latest: latestVersion },
-            )}
-          </p>
-        )}
-        {changed && (
-          <p className="mt-1 break-words text-[12px] leading-relaxed text-muted-foreground">{changed}</p>
-        )}
-        <p className="mt-1 text-[11.5px] text-faint">
-          {source === 'hosted-agent'
-            ? t('teamShare.skillConflictSourceHosted', '修改来源：本机 Hosted Agent')
-            : t('teamShare.skillConflictSourceMember', '修改来源：本机成员目录')}
-        </p>
-        <div className="mt-3 flex flex-wrap items-center gap-2">
-          <button
-            type="button"
-            onClick={onViewDiff}
-            disabled={busy}
-            className="text-[12px] text-muted-foreground underline-offset-2 hover:text-foreground hover:underline disabled:opacity-40"
-          >
-            {t('teamShare.skillConflictViewDiff', 'View changes')}
-          </button>
-          <span className="flex-1" />
-          {canPublish && !isStaleDirty && (
-            <Button
-              type="button"
-              onClick={onPublish}
-              disabled={busy}
-              className="h-8 gap-1.5 bg-coral text-[13px] font-semibold text-white hover:bg-coral/90"
-            >
-              {busy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Upload className="h-3.5 w-3.5" />}
-              {t('teamShare.skillConflictPublish', 'Publish as v{{v}}', {
-                v: (latestVersion ?? 0) + 1,
-              })}
-            </Button>
-          )}
-          {isStaleDirty && latestVersion != null && (
-            <Button
-              type="button"
-              variant="ghost"
-              onClick={onRebaseOnLatest}
-              disabled={busy}
-              className="h-8 gap-1.5 text-[13px] text-muted-foreground hover:text-foreground"
-            >
-              {t('teamShare.skillRebaseOnLatest', 'Apply team v{{v}}', { v: latestVersion })}
-            </Button>
-          )}
-          <Button
-            type="button"
-            variant="ghost"
-            onClick={onFork}
-            disabled={busy}
-            className="h-8 gap-1.5 text-[13px] text-muted-foreground hover:text-foreground"
-          >
-            <Copy className="h-3.5 w-3.5" />
-            {t('teamShare.skillConflictFork', 'Save as personal')}
-          </Button>
-          <Button
-            type="button"
-            variant="ghost"
-            onClick={onDiscard}
-            disabled={busy}
-            className="h-8 gap-1.5 text-[13px] text-muted-foreground hover:text-foreground"
-          >
-            <Trash2 className="h-3.5 w-3.5" />
-            {isStaleDirty
-              ? t('teamShare.skillConflictDiscardStale', 'Discard draft')
-              : t('teamShare.skillConflictDiscard', 'Discard local changes')}
-          </Button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function PublishVersionSheet({
-  item,
-  nextVersion,
-  baseVersion,
-  open,
-  busy,
-  changePreview,
-  onLoadDraftMetadata,
-  onClose,
-  onSubmit,
-}: {
-  item: TeamSkillItem
-  nextVersion: number
-  /** The version this machine's copy was built from. */
-  baseVersion: number | null
-  open: boolean
-  busy: boolean
-  changePreview?: { modified: string[]; deleted: string[]; added: string[] }
-  onLoadDraftMetadata: () => Promise<TeamSkillDraftMetadata>
-  onClose: () => void
-  onSubmit: (input: {
-    changelog: string
-    summary: string
-    category: TeamSkillCategory
-    whenToUse: string
-    whenNotToUse: string
-    requires: string[]
-  }) => Promise<void>
-}) {
-  const { t } = useTranslation()
-  const [changelog, setChangelog] = React.useState('')
-  const [summary, setSummary] = React.useState(item.summary ?? '')
-  const [category, setCategory] = React.useState<TeamSkillCategory>(
-    (TEAM_SKILL_CATEGORIES.includes(item.category as TeamSkillCategory)
-      ? item.category
-      : 'general') as TeamSkillCategory,
-  )
-  const [whenToUse, setWhenToUse] = React.useState(item.whenToUse ?? '')
-  const [whenNotToUse, setWhenNotToUse] = React.useState(item.whenNotToUse ?? '')
-  const [requiresText, setRequiresText] = React.useState((item.requires ?? []).join(', '))
-  const [metadataLoading, setMetadataLoading] = React.useState(false)
-  const [metadataReady, setMetadataReady] = React.useState(false)
-  const [metadataError, setMetadataError] = React.useState<string | null>(null)
-
-  React.useEffect(() => {
-    if (!open) return
-    setChangelog('')
-    setSummary(item.summary ?? '')
-    setCategory(
-      (TEAM_SKILL_CATEGORIES.includes(item.category as TeamSkillCategory)
-        ? item.category
-        : 'general') as TeamSkillCategory,
-    )
-    setWhenToUse(item.whenToUse ?? '')
-    setWhenNotToUse(item.whenNotToUse ?? '')
-    setRequiresText((item.requires ?? []).join(', '))
-    setMetadataLoading(true)
-    setMetadataReady(false)
-    setMetadataError(null)
-    void onLoadDraftMetadata()
-      .then((draft) => {
-        if (typeof draft.summary === 'string') setSummary(draft.summary)
-        if (
-          typeof draft.category === 'string' &&
-          TEAM_SKILL_CATEGORIES.includes(draft.category as TeamSkillCategory)
-        ) {
-          setCategory(draft.category as TeamSkillCategory)
-        }
-        if (typeof draft.whenToUse === 'string') setWhenToUse(draft.whenToUse)
-        if (typeof draft.whenNotToUse === 'string') setWhenNotToUse(draft.whenNotToUse)
-        if (draft.requires !== undefined) setRequiresText((draft.requires ?? []).join(', '))
-        setMetadataReady(true)
-      })
-      .catch((e) => {
-        setMetadataError(e instanceof Error ? e.message : String(e))
-      })
-      .finally(() => setMetadataLoading(false))
-  }, [open, item, onLoadDraftMetadata])
-
-  if (!open) return null
-
-  const field = 'w-full rounded-[8px] border border-border bg-background px-3 py-2 text-[13px] outline-none focus:border-coral/60'
-  const label = 'text-[11px] font-semibold uppercase tracking-wide text-faint'
-  const metaHint = (draftValue: string, registryValue: string | null | undefined) =>
-    draftValue.trim() !== (registryValue ?? '').trim() ? (
-      <span className="ml-2 font-normal normal-case text-faint">
-        {t('teamShare.skillPublishDraftDiff', 'differs from registry')}
-      </span>
-    ) : null
-
-  const previewCount =
-    (changePreview?.modified.length ?? 0) +
-    (changePreview?.deleted.length ?? 0) +
-    (changePreview?.added.length ?? 0)
-
-  // Publishing sends the directory as it stands, and this directory was built
-  // from `baseVersion`. If the team moved on while auto-follow was held back by
-  // the local edit, everything published in between is about to be replaced by
-  // content that predates it. That is sometimes exactly right and there is no
-  // merge to offer, but it must not happen silently — the diff button beside it
-  // deliberately compares against the installed version, so this warning is the
-  // only place the user learns those versions exist at all.
-  const latest = item.latestVersion ?? 0
-  const skipped = baseVersion != null && latest > baseVersion ? latest - baseVersion : 0
-  // Listed individually while a person can still read them as a list; a range
-  // past that, because "v2, v3, v4, …, v37" is not more informative than
-  // "v2–v37" and wraps the dialog.
-  const overwritten =
-    skipped === 0
-      ? ''
-      : skipped <= 3
-        ? Array.from({ length: skipped }, (_, i) => `v${(baseVersion as number) + i + 1}`).join(', ')
-        : `v${(baseVersion as number) + 1}–v${latest}`
-
-  return (
-    <ModalShell
-      title={t('teamShare.skillPublishTitle', 'Publish v{{v}}', { v: nextVersion })}
-      hint={t(
-        'teamShare.skillPublishHint',
-        'Publishes what is on disk right now. Everyone on the team moves to it automatically.',
-      )}
-      onClose={onClose}
-      footer={
-        <>
-          <Button type="button" variant="ghost" onClick={onClose} disabled={busy} className="h-8 text-[13px]">
-            {t('common.cancel', 'Cancel')}
-          </Button>
-          <Button
-            type="button"
-            disabled={
-              !changelog.trim() ||
-              !summary.trim() ||
-              busy ||
-              metadataLoading ||
-              !metadataReady ||
-              !!metadataError
-            }
-            onClick={() =>
-              void onSubmit({
-                changelog: changelog.trim(),
-                summary: summary.trim(),
-                category,
-                whenToUse: whenToUse.trim(),
-                whenNotToUse: whenNotToUse.trim(),
-                requires: requiresText
-                  .split(',')
-                  .map((part) => part.trim())
-                  .filter(Boolean),
-              })
-            }
-            className="h-8 gap-1.5 bg-coral text-[13px] font-semibold text-white hover:bg-coral/90 disabled:opacity-40"
-          >
-            {busy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Upload className="h-3.5 w-3.5" />}
-            {t('teamShare.skillPublishSubmit', 'Publish')}
-          </Button>
-        </>
-      }
-    >
-      {skipped > 0 && (
-        <div className="rounded-[8px] border border-border border-l-2 border-l-foreground bg-paper px-3 py-2">
-          <p className="text-[12px] font-semibold text-foreground">
-            {t('teamShare.skillPublishStaleTitle', 'Your copy is based on v{{base}}', {
-              base: baseVersion,
-            })}
-          </p>
-          <p className="mt-0.5 text-[12px] leading-relaxed text-muted-foreground">
-            {t(
-              'teamShare.skillPublishStaleBody',
-              '{{versions}} shipped after that. Publishing v{{next}} replaces them with your copy, so those changes will be gone.',
-              { versions: overwritten, next: nextVersion },
-            )}
-          </p>
-        </div>
-      )}
-      {previewCount > 0 && (
-        <div className="rounded-[8px] border border-border-soft bg-paper/60 px-3 py-2 text-[12px] text-muted-foreground">
-          {t('teamShare.skillPublishPreview', 'Publishing {{count}} local file change(s) from disk.', {
-            count: previewCount,
-          })}
-        </div>
-      )}
-      {metadataLoading && (
-        <div className="flex items-center gap-2 text-[12px] text-muted-foreground">
-          <Loader2 className="h-3.5 w-3.5 animate-spin" />
-          {t('teamShare.skillPublishLoadingDraft', 'Reading draft metadata…')}
-        </div>
-      )}
-      {metadataError && (
-        <div className="rounded-[8px] border border-border border-l-2 border-l-destructive bg-paper px-3 py-2 text-[12px] text-muted-foreground">
-          {t('teamShare.skillPublishDraftMetadataFailed', 'Could not read draft metadata: {{msg}}', {
-            msg: metadataError,
-          })}
-        </div>
-      )}
-      <label className="block space-y-1">
-        <span className={label}>{t('teamShare.skillShareChangelog', 'Changelog')}</span>
-        <textarea
-          value={changelog}
-          onChange={(e) => setChangelog(e.target.value)}
-          rows={2}
-          autoFocus
-          placeholder={t('teamShare.skillPublishChangelogHint', 'What changed, in one line')}
-          className={field}
-        />
-      </label>
-      <label className="block space-y-1">
-        <span className={label}>
-          {t('teamShare.skillShareSummary', 'Summary')}
-          {metaHint(summary, item.summary)}
-        </span>
-        <input value={summary} onChange={(e) => setSummary(e.target.value)} maxLength={200} className={field} />
-      </label>
-      <label className="block space-y-1">
-        <span className={label}>
-          {t('teamShare.skillShareCategory', 'Category')}
-          {metaHint(category, item.category)}
-        </span>
-        <Select value={category} onValueChange={(v) => setCategory(v as TeamSkillCategory)}>
-          <SelectTrigger className="h-auto w-full rounded-[8px] border-border bg-background px-3 py-2 text-[13px] shadow-none">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {TEAM_SKILL_CATEGORIES.map((c) => (
-              <SelectItem key={c} value={c} className="text-[13px]">
-                {c}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </label>
-      <label className="block space-y-1">
-        <span className={label}>
-          {t('teamShare.skillWhenToUse', 'When to use')}
-          {metaHint(whenToUse, item.whenToUse)}
-        </span>
-        <textarea value={whenToUse} onChange={(e) => setWhenToUse(e.target.value)} rows={3} className={field} />
-      </label>
-      <label className="block space-y-1">
-        <span className={label}>
-          {t('teamShare.skillWhenNotToUse', 'When not to use')}
-          {metaHint(whenNotToUse, item.whenNotToUse)}
-        </span>
-        <textarea
-          value={whenNotToUse}
-          onChange={(e) => setWhenNotToUse(e.target.value)}
-          rows={3}
-          className={field}
-        />
-      </label>
-      <label className="block space-y-1">
-        <span className={label}>
-          {t('teamShare.skillRequires', 'Requires')}
-          {metaHint(requiresText, (item.requires ?? []).join(', '))}
-        </span>
-        <input
-          value={requiresText}
-          onChange={(e) => setRequiresText(e.target.value)}
-          placeholder={t('teamShare.skillRequiresHint', 'comma-separated, e.g. git, jq')}
-          className={field}
-        />
-      </label>
-    </ModalShell>
-  )
-}
-
-function ForkSheet({
-  slug,
-  open,
-  busy,
-  onClose,
-  onSubmit,
-}: {
-  slug: string
-  open: boolean
-  busy: boolean
-  onClose: () => void
-  onSubmit: (newSlug: string) => Promise<void>
-}) {
-  const { t } = useTranslation()
-  const [newSlug, setNewSlug] = React.useState(`${slug}-mine`)
-
-  React.useEffect(() => {
-    if (open) setNewSlug(`${slug}-mine`)
-  }, [open, slug])
-
-  if (!open) return null
-
-  return (
-    <ModalShell
-      title={t('teamShare.skillForkTitle', 'Save as personal skill')}
-      // The rename is not a preference. Local skills outrank team skills in the
-      // loader, so a fork keeping the original slug would shadow the team copy
-      // and silently undo the auto-follow this action exists to preserve.
-      hint={t(
-        'teamShare.skillForkHint',
-        'Keeps your edits under a new name. The team version goes back to updating automatically.',
-      )}
-      onClose={onClose}
-      footer={
-        <>
-          <Button type="button" variant="ghost" onClick={onClose} disabled={busy} className="h-8 text-[13px]">
-            {t('common.cancel', 'Cancel')}
-          </Button>
-          <Button
-            type="button"
-            disabled={!newSlug.trim() || newSlug.trim() === slug || busy}
-            onClick={() => void onSubmit(newSlug.trim())}
-            className="h-8 gap-1.5 bg-coral text-[13px] font-semibold text-white hover:bg-coral/90 disabled:opacity-40"
-          >
-            {busy ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Copy className="h-3.5 w-3.5" />}
-            {t('teamShare.skillForkSubmit', 'Save a copy')}
-          </Button>
-        </>
-      }
-    >
-      <label className="block space-y-1">
-        <span className="text-[11px] font-semibold uppercase tracking-wide text-faint">
-          {t('teamShare.skillShareSlug', 'Slug')}
-        </span>
-        <input
-          value={newSlug}
-          onChange={(e) => setNewSlug(e.target.value)}
-          autoFocus
-          className="w-full rounded-[8px] border border-border bg-background px-3 py-2 font-mono text-[13px] outline-none focus:border-coral/60"
-        />
-      </label>
-    </ModalShell>
-  )
-}
-
-function DiffSheet({
-  slug,
-  diffs,
-  teamDiffs,
-  loading,
-  teamLoading,
-  showTeamTab,
-  diffTab,
-  onDiffTabChange,
-  isDark,
-  onClose,
-}: {
-  slug: string
-  diffs: TeamSkillFileDiff[] | null
-  teamDiffs: TeamSkillFileDiff[] | null
-  loading: boolean
-  teamLoading: boolean
-  showTeamTab: boolean
-  diffTab: 'local' | 'team'
-  onDiffTabChange: (tab: 'local' | 'team') => void
-  isDark: boolean
-  onClose: () => void
-}) {
-  const { t } = useTranslation()
-  const activeDiffs = diffTab === 'team' ? teamDiffs : diffs
-  const activeLoading = diffTab === 'team' ? teamLoading : loading
-  return (
-    <ModalShell
-      wide
-      title={t('teamShare.skillDiffTitle', 'Your changes to {{slug}}', { slug })}
-      hint={
-        diffTab === 'team'
-          ? t('teamShare.skillTeamDiffHint', 'What the team shipped while your draft was based on an older version.')
-          : t(
-              'teamShare.skillDiffHint',
-              'Compared against the version you installed, not the newest one.',
-            )
-      }
-      onClose={onClose}
-      footer={
-        <Button type="button" variant="ghost" onClick={onClose} className="h-8 text-[13px]">
-          {t('common.close', 'Close')}
-        </Button>
-      }
-    >
-      {showTeamTab && (
-        <div className="mb-3 flex gap-2">
-          <button
-            type="button"
-            onClick={() => onDiffTabChange('local')}
-            className={cn(
-              'rounded-[7px] px-2.5 py-1 text-[12px]',
-              diffTab === 'local'
-                ? 'bg-selected text-foreground'
-                : 'text-muted-foreground hover:text-foreground',
-            )}
-          >
-            {t('teamShare.skillDiffTabLocal', 'Your edits')}
-          </button>
-          <button
-            type="button"
-            onClick={() => onDiffTabChange('team')}
-            className={cn(
-              'rounded-[7px] px-2.5 py-1 text-[12px]',
-              diffTab === 'team'
-                ? 'bg-selected text-foreground'
-                : 'text-muted-foreground hover:text-foreground',
-            )}
-          >
-            {t('teamShare.skillDiffTabTeam', 'Team updates')}
-          </button>
-        </div>
-      )}
-      {activeLoading && (
-        <div className="flex items-center gap-2 text-[12px] text-muted-foreground">
-          <Loader2 className="h-3.5 w-3.5 animate-spin" />
-          {t('common.loading', 'Loading…')}
-        </div>
-      )}
-      {!activeLoading && activeDiffs?.length === 0 && (
-        <p className="text-[12px] text-muted-foreground">
-          {t('teamShare.skillDiffEmpty', 'Nothing to compare.')}
-        </p>
-      )}
-      {!activeLoading &&
-        activeDiffs?.map((d) => (
-          <div key={d.path} className="space-y-1">
-            <div className="font-mono text-[11px] text-faint">{d.path}</div>
-            {d.binary ? (
-              <p className="text-[12px] text-muted-foreground">
-                {t('teamShare.skillDiffBinary', 'Binary file — not shown.')}
-              </p>
-            ) : (
-              <Suspense
-                fallback={
-                  <div className="text-[12px] text-muted-foreground">{t('common.loading', 'Loading…')}</div>
-                }
-              >
-                <DiffRenderer
-                  before={d.baseline ?? ''}
-                  after={d.current ?? ''}
-                  filePath={d.path}
-                  isDark={isDark}
-                />
-              </Suspense>
-            )}
-          </div>
-        ))}
-    </ModalShell>
-  )
-}
 
 export function SkillDetail({ slug }: { slug: string }) {
   const { t } = useTranslation()
