+++ packages/app/src/components/teamshare/skill-detail/format.ts
@@ -1,0 +1,8 @@
+
+
+export function formatShortDate(iso: string | null | undefined): string | null {
+  if (!iso) return null
+  const d = new Date(iso)
+  if (Number.isNaN(d.getTime())) return null
+  return d.toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' })
+}
